refactor: PPS-41 refactored backends package organization: now backends are stored inside a separate `backends` package inside http
