+++ src/main/scala/utility/http/Client.scala
@@ -6,18 +6,40 @@ import utility.http.Configuration.ClientConfiguration
 import scala.concurrent.duration.Duration
 
 /**
+ * Core of the HTTP logic. A [[HttpClient]] can mix-in a Backend to gain the corresponding, underlying implementation of
+ * HTTP calls. Each Backend must specify the type of the generated HTTP responses: it can also (and in fact should) be a
+ * type that does not depend on the underlying HTTP library.
+ *
+ * Note that while the [[Request]] type is used inside all the backends, the response's one [[R]] is specific to the
+ * used Backend.
+ *
+ * @tparam R
+ *   type of the HTTP responses generated by the Backend
+ */
+trait Backend[R] extends HttpClient:
+  /**
+   * Sends the provided [[Request]] using the Backend implementation
+   *
+   * @param request
+   *   request to be sent
+   * @return
+   *   a response of type [[R]], depending on the Backend implementation
+   */
+  def send(request: Request): R
+
+/**
  * Empty trait that represents an HTTP client. To be useful, you need to mix-in [[Backend]] Traits
  */
 trait HttpClient(configuration: ClientConfiguration)
 
 object Clients:
-  import utility.http.Backends.SttpBackend
+  import utility.http.backends.SttpSyncBackend
 
   /**
    * Simple HTTP client used for synchronous HTTP calls with the sttp library as backend
    */
   class SimpleHttpClient(configuration: ClientConfiguration = Configuration.default)
-      extends HttpClient(configuration) with SttpBackend
+      extends HttpClient(configuration) with SttpSyncBackend
 
 type Client[R] = HttpClient & Backend[R]
